rename entity folder to entities
